Mark which tags map to base Python types
The main point of this commit is to mark which tags have payloads that
map to base Python types like int, str, float, etc. New methods are
added for each class with a "primitive" payload for deserialization and
serialization. This is useful for testing and code-reuse.

Also introduce a small bit of overhead to optimize TAG_List in the case
of a primitive being used. In practice, TAG_Int_Array, etc would be used
for numerics and so this only makes sense for lists of TAG_String. Since
the only test data I have is for players, villages, and mineshafts,
there's no evidence the optimization was necessary. I'll revisit in
later optimization passes with more map data.