Include memory and line count profiling for the previous commit
This used a locally modified version of `memory_profiler` (below) to
produce a broader range of stats about memory and also the number of
times each line was executed.

https://github.com/fabianp/memory_profiler

A comparison of the two files shows that no effective optimization
happened due to the last commit. The data is included for future
reference during profiling experiments.